+++ components/client/src/request.rs
@@ -117,7 +117,7 @@ pub enum Headers {
         stream_id: u64,
         replica_count: Option<u8>,
         ack_count: Option<u8>,
-        epoch: Option<u8>,
+        epoch: Option<u64>,
     },
 }
